@@ -22,7 +22,7 @@
   "license": "MIT",
   "type": "project",
   "require": {
-    "php": "^7.2.5",
+    "php": "^7.4",
     "ext-intl": "*",
     "ext-json": "*",
     "ext-pdo": "*",
@@ -46,7 +46,7 @@
     "gmsantos/inspiring": "^3.0",
     "league/flysystem": "^1.0.66",
     "mockery/mockery": "^1.3.1",
-    "phpunit/phpunit": "^8.5.3"
+    "phpunit/phpunit": "^9.2.6"
   },
   "autoload": {
     "classmap": [
